Refactoring and adjusting color scheme
